requirements-phase2.txt
# === Phase 0+1: Foundation ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
starlette>=0.38.0
python-dotenv>=1.0.0
pydantic>=2.7.0
structlog>=24.1.0
PyYAML>=6.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
jinja2>=3.1.0
httpx>=0.27.0

# === Phase 2: Discovery Engine ===
playwright>=1.44.0

# === Dev / Testing ===
# pip install -r requirements-dev.txt for these
